project_boilerplate/app/requirements/base.txt
cx-Oracle==5.2.1
marshmallow==2.7.1
pymssql==2.1.2
restless==2.0.1
SQLAlchemy==1.0.12
tornado==4.3
python_dotenv=0.5.0
